+++ public/index.php
@@ -3,9 +3,5 @@
 require dirname(__DIR__) . '/vendor/autoload.php';
 
 use App\Application;
-use Database\Seed;
 
 $app = new Application();
-$app->run();
-$seed = new Seed();
-$seed->run();